Added the functionality to use in diff machines via local network

+++ src/Contexts/Notes/NotesContext.jsx
@@ -2,7 +2,7 @@ import { createContext, useState } from "react";
 import axios from "axios";
 
 export const NoteContext = createContext();
-const host = "http://localhost:4000";
+const host = "http://192.168.0.143:4000";
 const authToken =
   "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0ODYwODc4YWNjMWEwMmZhNzViYWE3NiIsImlhdCI6MTY4NzQzNDg3MH0.J6yOMNXlXk2ng3G5iAfd8hhg-Hj0qld-Ibo87QApais";
 
@@ -26,21 +26,29 @@ export const NoteContextProvider = ({ children }) => {
   };
 
   const updateNote = async (noteId, updatedNotes) => {
-
+    if (!noteId) {
+      return console.log("Provide a valid note!");
+    }
+      if (!updatedNotes) {
+        return console.log("Provide a valid update note!")
+      }
     try {
-        const response = await axios.patch(`${host}/api/notes/updatenote/${noteId}`,updatedNotes, {
+         await axios.patch(`${host}/api/notes/updatenote/${noteId}`,updatedNotes, {
             headers:{
                 'Content-Type':'application/json',
                 'auth-token':authToken
             },
             
         });
-        setNotes([...notes, response.data]);
+        
+        setNotes([...notes]);
     } catch (error) {
         console.error(error);  
     }
   };
   const deleteNote = async (noteId) => {
+    if(!noteId) return console.log("Please provide a VALID ID");
+    if(!notes.find((note)=>note._id === noteId)) return console.log("Invalid Note to delete");
     try {
          await axios.delete(`${host}/api/notes/deletenote/${noteId}`, {
             headers:{
